refactor: move runtime dev logic to plugin
